% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utiles-clean.R
\name{clean_lsat}
\alias{clean_lsat}
\title{=====' reflects that lsat was taken multiple times}
\usage{
clean_lsat(text_extract_lsat, lsacNum)
}
\description{
=====' reflects that lsat was taken multiple times
}
