% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-core.R
\name{convert_df}
\alias{convert_df}
\title{Convert object from `extracted_areas` to data frames}
\usage{
convert_df(text_extract)
}
\arguments{
\item{text_extract}{An object from `extracted_areas`}
}
\value{
A list with each element of the list representing a data frame with
     information from the scraped LSAC report.
}
\description{
This function takes as inputs an object created with `extracted_areas` and
converts the object's data to a series of dataframes.  The data frames are
output in a list.
}
