% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-core.R
\name{extracted_areas}
\alias{extracted_areas}
\title{Extract text from LSAC reports}
\usage{
extracted_areas(file_name, page_num)
}
\arguments{
\item{file_name}{String.  Full path to the PDF file to be scraped.}

\item{page_num}{Integer.  The page number of the PDF file to be scraped.}
}
\value{
The raw, untransformed and uncleaned, data from the LSAC report; returned
     as data frames stored within a list.
}
\description{
This function extracts the raw text from the LSAC reports.
The extracted text is unformatted and each extracted section is
returned in a list.
}
