% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-core.R
\name{text_to_list}
\alias{text_to_list}
\title{Extract text areas and place in individual dataframes}
\usage{
text_to_list(pdf_file)
}
\arguments{
\item{pdf_file}{String.  Full path to the PDF file to be scraped.}
}
\value{
A nested list with each student as the highest hierarchy in the list, and
     each student is another list with a different data frame for each table.
}
\description{
This function extracts and cleans the data and places it into
data frames for individual students.  Each student's data will be placed
in a separate data frame.
}
