% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-core.R
\name{varFilter}
\alias{varFilter}
\title{Filter out rows in the transcript with no values}
\usage{
varFilter(df, columns)
}
\arguments{
\item{df}{Data frame to filter rows}

\item{columns}{Columns to filter}
}
\description{
Filter out rows in the transcript with no values
}
